Harden defenses against database failures (OCECDR-5203)
In the scheduled job to refresh the glossifier data, when the
database communication fails, we must make sure we don't leave
behind partially populated (or completely unpopulated) instance
properties. So instead of starting the property method with
`self._property_name = {}` (or similar code) use a local variable
to hold the data as we accumulate it, and only at the end of the
method when we're satisfied that the population of that variable
has completely succeeded do we assign it to the object's property.